Refactored the frontend classes.
